Add demo mode
This mode starts when the number of WiFi connection attemps in
MAX_TRIES is exceeded. After that the Wifi / MQTT connection code is
skipped and the PM value is updated according to demoAction().

demoAction injects PM values into the code to simulate changing
PM values from the ESDK.

Once in demo mode reset is required to restart normally.